+++ tests/unit/test_healthcheck.py
@@ -86,5 +86,33 @@ class BasicHealthCheckTest(unittest.TestCase):
         jr = json.loads(message)
         self.assertEqual("My custom section", jr["custom_section"])
 
+
+class TimeoutHealthCheckTest(unittest.TestCase):
+
+    def test_default_timeout_function_should_failing_check(self):
+        def timeout_check():
+            import time
+            time.sleep(15)
+            return True, "Waited for 10 seconds"
+
+        hc = HealthCheck(checkers=[timeout_check])
+        message, status, headers = hc.run()
+        self.assertEqual(500, status)
+        jr = json.loads(message)
+        self.assertEqual("failure", jr["status"])
+
+    def test_error_timeout_function_should_failing_check(self):
+        def timeout_check():
+            import time
+            time.sleep(5)
+            return True, "Waited for 10 seconds"
+
+        hc = HealthCheck(checkers=[timeout_check], error_timeout=2)
+        message, status, headers = hc.run()
+        self.assertEqual(500, status)
+        jr = json.loads(message)
+        self.assertEqual("failure", jr["status"])
+
+
 if __name__ == '__main__':
     unittest.main()