+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Healthcheck Changelog
 
+### Next Release
+- Adds timeout on execution checkers.
+
 ### 1.5.0
 - Adds `HealthCheck.add_section` to increase output